backend/pyproject.toml
[project]
name = "dragonweaver-backend"
version = "0.1.0"
description = "DragonWeaver backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.32,<3.0.0",
  "greenlet>=3.0.0,<4.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "pwdlib[argon2]>=0.2.0,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.11.0,<2.0.0",
  "types-python-jose>=3.3.4.20240106,<4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
